fix cast to int

+++ champions/src/main/java/me/mykindos/betterpvp/champions/champions/skills/skills/gladiator/axe/SpiritOfTheBear.java
@@ -43,7 +43,7 @@ public class SpiritOfTheBear extends Skill implements InteractSkill, CooldownSki
                 "",
                 "Call upon the spirit of the bear,",
                 "granting all allies within <val>" + (radius + (level)) + "</val> blocks",
-                "<effect>Resistance II</effect< for <stat>" + (int)duration + "</stat> seconds.",
+                "<effect>Resistance II</effect< for <stat>" + duration + "</stat> seconds.",
                 "",
                 "Cooldown: <val>" + getCooldown(level)
         };
